Work around className issue
Both Elm and components would like to add to a node's CSS classes. This
is problematic, since Elm assumes full control over the `className`
property. We utilize the fact that MDC only every utilises the
`classList` property. This way, we can override the `className` setter
to mitigate between Elm's and MDC's actions on CSS classes at the cost
of a performance penalty.